Refactor: Consolidate configuration management and streamline reset logic
- Created a new `configurations.js` file to centralize the creation of kit, shop, and rank config managers.
- Refactored the `resetConfigSection` function in `configManager.js` to use a registry lookup instead of a chain of if-statements.
- Updated all imports across the codebase to use the new `configurations.js` file.
- Deleted the redundant `kitsConfigManager.js`, `shopConfigManager.js`, and `ranksConfigManager.js` files.
